# backend/src/swagger.yaml
openapi: "3.0.3"
info:
  title: HalalChain V2 API
  description: Enterprise Halal Supply Chain Platform — REST API
  version: "2.0.0"
  contact:
    name: HalalChain Team
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.halalchain.app
    description: Production server

# ─────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token

  # ─────────────────────────────────────────────────────────────
  # Schemas
  # ─────────────────────────────────────────────────────────────
  schemas:
...
    HalalCertificate:
      type: object
      required:
        - id
        - supplierId
        - certificateNumber
        - issuedBy
        - issueDate
        - expiryDate
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "880e8400-e29b-41d4-a716-446655440003"
        supplierId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        certificateNumber:
          type: string
          example: "JAKIM/HPC/2024/001234"
        issuedBy:
          type: string
          example: "JAKIM"
        issueDate:
          type: string
          format: date
          example: "2024-01-01"
        expiryDate:
          type: string
          format: date
          example: "2025-12-31"
        fileUrl:
          type: string
          nullable: true
          example: "https://storage.halalchain.app/certs/cert-001.pdf"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-02T07:00:00.000Z"

    # ── Inventory ──────────────────────────────────────────────
...
    PaginatedResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
        - totalPages
      properties:
        data:
          type: array
          items:
            type: object
          description: Array of result items for the current page
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 85
        totalPages:
          type: integer
          minimum: 0
          example: 5

    # ── ErrorResponse ──────────────────────────────────────────
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Resource not found"

# ─────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────
paths:
...
  /api/certificates:
    get:
      tags:
        - Certificates
      summary: List halal certificates (paginated, filterable)
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: certificateNumber
          in: query
          description: Filter by certificate number (case-insensitive, partial match)
          schema:
            type: string
            example: "JAKIM"
        - name: expiryBefore
          in: query
          description: Return certificates expiring on or before this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2025-12-31"
        - name: expiryAfter
          in: query
          description: Return certificates expiring on or after this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2024-01-01"
      responses:
        "200":
          description: Paginated list of halal certificates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/HalalCertificate"
        "400":
          description: Invalid date format for expiryBefore or expiryAfter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid date format. Expected YYYY-MM-DD."
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Certificates
      summary: Create a new halal certificate (ADMIN only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - supplierId
                - certificateNumber
                - issuedBy
                - issueDate
                - expiryDate
              properties:
                supplierId:
                  type: string
                  format: uuid
                  example: "660e8400-e29b-41d4-a716-446655440001"
                certificateNumber:
                  type: string
                  minLength: 3
                  maxLength: 100
                  example: "JAKIM/HPC/2024/001234"
                issuedBy:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "JAKIM"
                issueDate:
                  type: string
                  format: date
                  example: "2024-01-01"
                expiryDate:
                  type: string
                  format: date
                  example: "2025-12-31"
                fileUrl:
                  type: string
                  format: uri
                  example: "https://storage.halalchain.app/certs/cert-001.pdf"
      responses:
        "201":
          description: Certificate created
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificate:
                    $ref: "#/components/schemas/HalalCertificate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/certificates/{id}:
    get:
      tags:
        - Certificates
      summary: Get a single certificate by ID
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Certificate details
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificate:
                    $ref: "#/components/schemas/HalalCertificate"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Certificates
      summary: Update a certificate (ADMIN only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                certificateNumber:
                  type: string
                  minLength: 3
                  maxLength: 100
                issuedBy:
                  type: string
                  minLength: 2
                  maxLength: 100
                issueDate:
                  type: string
                  format: date
                expiryDate:
                  type: string
                  format: date
                fileUrl:
                  type: string
                  format: uri
      responses:
        "200":
          description: Certificate updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificate:
                    $ref: "#/components/schemas/HalalCertificate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Certificates
      summary: Delete a certificate (ADMIN only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Certificate deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Certificate deleted"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ══════════════════════════════════════════════════════════
  # Inventory
  # ══════════════════════════════════════════════════════════
...